Universal Monster Drop Compatibility
Changed the patch to monster drop tables to use test patches to see if "fill" exist or not before making the "fill" in the table. If "fill" isn't found, it makes "fill". If "fill" already exists, adds storage matter table to the end of the list.

This solves any incompatibility problems with mods loaded before Enhanced Storage that alter monster drops. (This was once a problem with the lucario once the lucario race mod was added to ES's "includes", where ES was overriding Essence of Aura drops, problem is non-existent now since how to obtain EoS was changed, but this mod needed this compatibility change anyways for any other mods that may have problems.)